LayoutTests/fast/webgpu/regression/repro_292315-expected.txt

CONSOLE MESSAGE: no validation error
